+++ extensions/vscode-rax-props/src/getBabelParserPlugins.js
@@ -1,0 +1,22 @@
+module.exports = function getBabelParserPlugins(language) {
+  const plugins = [
+    'doExpressions',
+    'objectRestSpread',
+    'decorators-legacy',
+    'classProperties',
+    'exportExtensions',
+    'asyncGenerators',
+    'functionBind',
+    'functionSent',
+    'dynamicImport'
+  ];
+
+  if (language === 'ts') {
+    plugins.unshift('typescript');
+  } else {
+    plugins.unshift('flow');
+    plugins.unshift('jsx');
+  }
+
+  return plugins;
+};
